# openapi.yaml
openapi: 3.0.0
info:
  title: 'Delilah Resto '
  description: API for delicious food orders. As a customer you can register, see the list of our products and place an order.Restaurant managers have the ability to receive orders and update them..
  termsOfService: https://github.com/hmfarias/03-acamica-delilah#readme
  contact:
    name: Marcelo Farias
    email: hmfarias7@gmail.com
  version: 1.0.0
servers:

  - description: Local Server
    url: http://localhost:3000/api/v1
...
paths:
...
  /roles:
    # ----------------------------------------------POST NEW ROLE
    post:
      tags:
        - Roles
      security: 
        - bearerAuth: 
          - read      
      summary: "Add a new user role to Delilah Resto's system"
      description: 'Admin level required'
      operationId: newRole
      requestBody:
        description: 'Required to post a new role'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/roleReq'
        required: true
      responses:
        '200':
          description: 'Successful operation - returns the added role'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/roleResp200'      
        '400':
          description: 'Some required fields is not present'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error400'
        '403':
          description: 'Admin Level required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error403'                   
        '409':
          description: 'Role already exists'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error409'           
        '500':
          description: 'The role could not be registered'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error500'                   
        'default':
          description: 'Internal error - Try again later'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorDefault'    
    # --------------------------------------------GET ALL ROLES            
    get:
      tags:
        - Roles
      security: 
        - bearerAuth: 
          - read      
      summary: 'List all roles'
      description: 'Admin level required'
      operationId: getRoles
      responses:
        '200':
          description: 'Successful operation - returns an array of roles'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/rolesResp200'          
        '403':
          description: 'Admin Level required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error403'                   
        '404':
          description: 'There are no roles in the database'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error404'                   
        'default':
          description: 'Internal error - Try again later'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorDefault'           
  /roles/{id}:
    # --------------------------------------------GET ROLE BY ID
    get:
      tags:
        - Roles
      security: 
        - bearerAuth: 
          - read
      summary: 'Find role by Id - Return a single role'
      description: 'Admin level required'
      operationId: getRole
      parameters:
        - name: id
          in: path
          description: 'ID of role to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Successful operation - returns the role with the Id specified by parameter'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/roleResp200'       
        '403':
          description: 'Admin Level required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error403'                   
        '404':
          description: 'The role not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error404'                   
        '410':
          description: 'The role is already deleted (soft deleted)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error410'                   
        'default':
          description: 'Internal error - Try again later'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorDefault'       
  /roles/update/{id}:
    # -------------------------------------------PUT UPDATE ROLE          
    put:
      tags:
        - Roles
      security: 
        - bearerAuth: 
          - read
      summary: 'Update an existent role'
      description: 'Admin level required'
      operationId: updateRole
      parameters:
        - name: id
          in: path
          description: 'ID of role that needs to be updated'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: 'Required to update a role'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/roleReqUpdate'
        required: true
      responses:
        '200':
          description: 'Successful operation - returns the updated role'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/roleResp200'          
        '400':
          description: 'Id parameter or name field are not present'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error400Param'      
        '403':
          description: 'Admin Level required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error403'                   
        '404':
          description: 'Role not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error404'       
        '500':
          description: 'The role could not be updated'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error500'                                   
        'default':
          description: 'Unexpected error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorDefault'       
  /roles/delete/{id}:
      # ---------------------------------------DELETE ROLE BY ID
    delete:
      tags:
        - Roles
      security: 
        - bearerAuth: 
          - read      
      summary: 'Deletes a role by Id'
      description: 'Admin level required'
      operationId: deleteRole
      parameters:
        - name: id
          in: path
          description: 'ID of role that needs to be deleted'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Successful operation - returns the deleted role'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/roleResp200'          
        '400':
          description: 'The Id parameter is required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error400Param'      
        '403':
          description: 'Admin Level required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error403'                   
        '404':
          description: 'The role not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error404'                     
        '410':
          description: 'The role is already deleted (soft deleted)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error410'     
        '500':
          description: 'The role could not be deleted'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error500'                           
        'default':
          description: 'Internal error - Try again later'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorDefault'    
  /roles/restore/{id}:
    # ----------------------------------------RESTORE ROLE BY ID
    put:
      tags:
        - Roles
      security: 
        - bearerAuth: 
          - read      
      summary: 'Restore a role by Id'
      description: 'Admin Level required'
      operationId: restoreRole
      parameters:
        - name: id
          in: path
          description: 'ID of role that needs to be restored'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Successful operation - returns the restored role'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/roleResp200'          
        '400':
          description: 'The Id parameter is required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error400Param'      
        '403':
          description: 'Admin Level required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error403'                   
        '404':
          description: 'The role not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error404'                     
        '406':
          description: 'The role is not deleted'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error406'                     
        '410':
          description: 'The role is deleted (soft deleted)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error410'     
        '500':
          description: 'The role could not be deleted'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error500'                           
        'default':
          description: 'Internal error - Try again later'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorDefault'        

  # ------------------------------------------------------------PATHS FOR USERS 
...
components:
...
  securitySchemes:
    bearerAuth:            
      type: http
      scheme: bearer
      bearerFormat: JWT

  #  --------------------------------------------------------SCHEMAS       
  schemas:
...
    role:
      required:
        - id
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 7
        name:
          type: string
          example: 'user'
        createdAt:
          type: string
          example: '|creation date and time|'
        updatedAt:
          type: string
          example: '|update date and time|'
        deletedAt:
          type: string
          example: '|null| or |delete date and time|'
...
    roles:
      type: array
      items: 
        $ref: '#/components/schemas/role'          
    roleReq:
      required:
        - name
      type: object
      properties:
        name:
          type: string
          example: 'user'
    roleReqUpdate:
      type: object
      properties:
        name:
          type: string
          example: 'user'
    roleResp200:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        data:
          type: object
          properties:
            role:
              $ref: '#/components/schemas/role'      
        message:
          $ref: '#/components/schemas/messageSuccess'
    rolesResp200:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        data:
          type: object
          properties:
            roles:
              $ref: '#/components/schemas/roles'     
        message:
          type: string
          example: 'Successfully recovered Roles'
          
      #  ----------------------------------SCHEMAS FOR LOGIN          
...
    messageSuccess:
      type: string
      example: 'Successful operation for |entity| ID: |entity ID|, Name: |entity name|'     
          
    #  ----------------------------------SCHEMAS FOR ERRORS
    errorDefault:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
          properties:
            error:
              type: string
              example: '|Error|'
        message:
          type: string
          example: 'Internal error - Try again later'
    error400:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
        message:
          type: string
          example: '|field| field is required |'  
    error400Param:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
        message:
          type: string
          example: '|parameter or field| |parameter or field name| is required | or has an incorrect format'  
    error403:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
        message:
          type: string
          example: 'Admin level required'  
...
    error404:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
        message:
          type: string
          example: '|entity| not found'     
    error406:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
        message:
          type: string
          example: '|entity| is not deleted'     
    error409:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
        message:
          type: string
          example: '|entity| |entity name| already exists'     
    error410:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
        message:
          type: string
          example: '|entity| is deleted - (soft deleted)'
    error500:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
        message:
          type: string
          example: 'Unexpected error - |message|'